wip(build): Migrate build and add tsconfig
