+++ .eslintrc
@@ -1,8 +1,9 @@
 {
   "root": true,
-  "parser": "babel-eslint",
+  "parser": "@babel/eslint-parser",
   "parserOptions": {
     "ecmaVersion": 11,
+    "requireConfigFile": false,
     "sourceType": "module",
     "ecmaFeatures": {
       "arrowFunctions": true,
