+++ .storybook/main.ts
@@ -1,13 +1,12 @@
-module.exports = {
-  stories: ['../stories/**/*.stories.@(js|jsx|mdx)'],
+import type { StorybookConfig } from '@storybook/react-webpack5'
+
+const config: StorybookConfig = {
+  stories: ['../stories/**/*.stories.@(js|jsx|mdx|tsx|ts)'],
   addons: [
     '@storybook/addon-links',
     '@storybook/addon-essentials',
     '@storybook/addon-mdx-gfm'
   ],
-  features: {
-    postcss: false
-  },
   framework: {
     name: '@storybook/react-webpack5',
     options: {}
@@ -16,3 +15,5 @@ module.exports = {
     autodocs: true
   }
 }
+
+export default config
