+++ package.json
@@ -5,6 +5,7 @@
   "main": "lib/index.js",
   "module": "es/index.js",
   "source": "src/index.js",
+  "type": "module",
   "files": [
     "css",
     "es",
@@ -12,7 +13,7 @@
     "umd"
   ],
   "scripts": {
-    "build": "node build",
+    "build": "ts-node --esm build.ts & tsc -noEmit --jsx react",
     "clean": "rm -rf lib es",
     "start": "BROWSER=none node build --serve",
     "format": "prettier-eslint \"$PWD/src/**/*.{js,jsx}\" --write",
@@ -32,6 +33,7 @@
     "styled-components": "^6.1.1"
   },
   "dependencies": {
+    "@babel/eslint-parser": "^7.23.3",
     "@emotion/is-prop-valid": "^1.2.1",
     "@mdi/js": "^5.8.55",
     "@mdi/react": "^1.4.0",
@@ -45,12 +47,14 @@
     "lodash": "^4.17.20",
     "prop-types": "^15.8.1",
     "styled-system": "^5.1.5",
+    "typescript": "^5.3.3",
     "use-resize-observer": "^9.1.0",
     "uuid": "^8.3.2"
   },
   "devDependencies": {
     "@babel/cli": "^7.19.3",
     "@babel/core": "^7.20.2",
+    "@babel/plugin-proposal-private-property-in-object": "^7.21.11",
     "@babel/preset-env": "^7.20.2",
     "@babel/preset-react": "^7.18.6",
     "@babel/register": "^7.18.9",
@@ -61,7 +65,22 @@
     "@storybook/react": "^7.0.26",
     "@storybook/react-webpack5": "^7.0.26",
     "@trivago/prettier-plugin-sort-imports": "^4.1.1",
-    "babel-eslint": "^10.1.0",
+    "@types/babel__core": "^7.20.5",
+    "@types/babel__preset-env": "^7.9.6",
+    "@types/babel__register": "^7.17.3",
+    "@types/eslint": "^8.56.0",
+    "@types/lint-staged": "^13.3.0",
+    "@types/lodash": "^4.14.202",
+    "@types/mocha": "^10.0.6",
+    "@types/node": "^20.10.5",
+    "@types/prop-types": "^15.7.11",
+    "@types/react": "^18.2.45",
+    "@types/react-dom": "^18.2.18",
+    "@types/styled-components": "^5.1.34",
+    "@types/styled-system": "^5.1.22",
+    "@types/styled-system__css": "^5.0.21",
+    "@types/styled-system__theme-get": "^5.0.4",
+    "@types/uuid": "^8.3.4",
     "babel-loader": "^8.1.0",
     "browserslist": "^4.21.4",
     "buffer": "^5.7.1",
@@ -69,6 +88,7 @@
     "esbuild-node-externals": "^1.11.0",
     "esbuild-plugin-browserslist": "^0.10.0",
     "esbuild-plugin-svgr": "^2.1.0",
+    "esbuild-serve": "^1.0.1",
     "eslint-config-react-app": "^7.0.1",
     "eslint-plugin-babel": "^5.3.0",
     "eslint-plugin-flowtype": "^4.7.0",
@@ -90,7 +110,8 @@
     "stylelint": "^15.10.3",
     "stylelint-config-recommended": "^3.0.0",
     "stylelint-config-styled-components": "^0.1.1",
-    "stylelint-processor-styled-components": "^1.10.0"
+    "stylelint-processor-styled-components": "^1.10.0",
+    "ts-node": "^10.9.2"
   },
   "author": "Mihai Ro",
   "homepage": "https://github.com/etvascom/ui",
