--- src/index.js
@@ -1,43 +1,0 @@
-import styled from 'styled-components'
-
-export * from './providers'
-export * from './utils'
-export * from './i18n'
-
-export * from './Button'
-export * from './Card'
-export * from './Dropdown'
-export * from './Input'
-export * from './TextArea'
-export * from './Typography'
-export * from './Tooltip'
-export * from './Grid'
-export * from './Table'
-export * from './Modal'
-export * from './Radio'
-export * from './Rating'
-export * from './NavBar'
-export * from './Icon'
-export * from './Flag'
-export * from './Checkbox'
-export * from './Form'
-export * from './Link'
-export * from './Skeleton'
-export * from './ErrorPages'
-export * from './EmbededApp'
-export * from './IntegrationHeader'
-export * from './Steps'
-export * from './Chip'
-
-export * from './SubdomainInput'
-export * from './Autocomplete'
-export * from './PhoneNumberInput'
-
-export * from './Box'
-export * from './Flex'
-export * from './Space'
-export * from './ActivityIndicator'
-export * from './Image'
-export * from './Touchable'
-
-export { styled }
